+++ packages/cryptoplease/lib/data/db/db.dart
@@ -1,7 +1,7 @@
 import 'package:drift/drift.dart';
 import 'package:injectable/injectable.dart';
 
-import '../../features/incoming_split_key_payments/bl/iskp_repository.dart';
+import '../../features/incoming_split_key_payments/module.dart';
 import '../../features/outgoing_direct_payments/bl/repository.dart';
 import '../../features/outgoing_split_key_payments/bl/repository.dart';
 import '../../features/payment_request/bl/repository.dart';
