+++ packages/cryptoplease/lib/features/incoming_split_key_payments/src/bl/pending_iskp_repository.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:injectable/injectable.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../../../core/split_key_payments.dart';
+import '../../../../core/split_key_payments.dart';
 
 @injectable
 class PendingISKPRepository {
