+++ packages/cryptoplease/lib/features/incoming_split_key_payments/src/presentation/first_part_ready_screen.dart
@@ -6,15 +6,15 @@ import 'package:solana/encoder.dart';
 import 'package:solana/solana.dart';
 import 'package:uuid/uuid.dart';
 
-import '../../../app/routes.dart';
-import '../../../core/dynamic_links_notifier.dart';
-import '../../../core/split_key_payments.dart';
-import '../../../di.dart';
-import '../../../gen/assets.gen.dart';
-import '../../../l10n/l10n.dart';
-import '../../../ui/button.dart';
-import '../../../ui/colors.dart';
-import '../../../ui/theme.dart';
+import '../../../../app/routes.dart';
+import '../../../../core/dynamic_links_notifier.dart';
+import '../../../../core/split_key_payments.dart';
+import '../../../../di.dart';
+import '../../../../gen/assets.gen.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/button.dart';
+import '../../../../ui/colors.dart';
+import '../../../../ui/theme.dart';
 import '../bl/iskp_bloc.dart';
 import '../bl/pending_iskp_repository.dart';
 import 'components/cancel_dialog.dart';
